* Fea #60, 添加Tokio支持   - 添加 SetFileCompletionNotificationModes   - 添加 GetQueuedCompletionStatusEx   - 添加 NtCancelIoFileEx

+++ src/YY-Thunks.UnitTest/api-ms-win-core-fibers.UnitTest.cpp
@@ -65,7 +65,7 @@ namespace api_ms_win_core_fibers
 
 				for (auto hHandle : hHandles)
 				{
-					Assert::AreEqual(WaitForSingleObject(hHandle, 1000), WAIT_OBJECT_0);
+					Assert::AreEqual(WaitForSingleObject(hHandle, 1000), (DWORD)WAIT_OBJECT_0);
 
 					CloseHandle(hHandle);
 				}
@@ -158,7 +158,7 @@ namespace api_ms_win_core_fibers
 
 				for (auto hHandle : hHandles)
 				{
-					Assert::AreEqual(WaitForSingleObject(hHandle, 1000), WAIT_OBJECT_0);
+					Assert::AreEqual(WaitForSingleObject(hHandle, 1000), (DWORD)WAIT_OBJECT_0);
 
 					CloseHandle(hHandle);
 				}
